Add: concatenate and use all available root certs
Autodownloaded certificates are IGNORED for the purposes of running
Demos which are configured at build system time. The certificate listed
in the documentation may be incorrect for some of the demos. It's easier
on the user to simply concatenate all certificates into a single chain
and let OpenSSL pick the correct one for the demo.